mem: A simple two-level architecture
I implement this for a experiment control. It consists of a simple direct-mapping SRAM cache and a
STT-RAM memory. Both latency parameters of medium are modeled from NVSim. In order to support LL/
SC instructions in ARM, i implemented a lockList in cache. In addition, a simple memory delay com
-ponent is used to simulate the differentiated read/write latency of STTRAM.

Change-Id: I474aa15184c5d9c8768068ea43d65905e9b03a75